@@ -1,29 +1,30 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
-        <LangVersion>8</LangVersion>
-        <Nullable>enable</Nullable>
-        <PackAsTool>true</PackAsTool>
-        <ToolCommandName>DcmAnonymize</ToolCommandName>
-        <Version>1.4.0</Version>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <RepositoryUrl>https://github.com/amoerie/DcmAnonymize</RepositoryUrl>
-        <Authors>Alexander Moerman</Authors>
-    </PropertyGroup>
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net6.0</TargetFramework>
+    <LangVersion>latest</LangVersion>
+    <Nullable>enable</Nullable>
+    <PackAsTool>true</PackAsTool>
+    <ToolCommandName>DcmAnonymize</ToolCommandName>
+    <Version>1.7.0</Version>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <RepositoryUrl>https://github.com/amoerie/DcmAnonymize</RepositoryUrl>
+    <Authors>Alexander Moerman</Authors>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+  </PropertyGroup>
 
-    <ItemGroup>
-      <PackageReference Include="CommandLineParser" Version="2.8.0" />
-      <PackageReference Include="fo-dicom.NetCore" Version="4.0.6" />
-      <PackageReference Include="KeyedSemaphores" Version="0.1.2" />
-    </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="CommandLineParser" Version="2.9.1" />
+    <PackageReference Include="fo-dicom" Version="5.0.3" />
+    <PackageReference Include="KeyedSemaphores" Version="4.1.0" />
+  </ItemGroup>
 
-    <ItemGroup>
-      <None Remove="Names\first-names.txt" />
-      <EmbeddedResource Include="Names\first-names.txt" />
-      <None Remove="Names\last-names.txt" />
-      <EmbeddedResource Include="Names\last-names.txt" />
-    </ItemGroup>
+  <ItemGroup>
+    <None Remove="Names\first-names.txt" />
+    <EmbeddedResource Include="Names\first-names.txt" />
+    <None Remove="Names\last-names.txt" />
+    <EmbeddedResource Include="Names\last-names.txt" />
+  </ItemGroup>
 
 </Project>
