@@ -1,8 +1,3 @@
-namespace DcmAnonymize.Names
-{
-    public class RandomName
-    {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-    }
-}
+namespace DcmAnonymize.Names;
+
+public record RandomName(string FirstName, string LastName);
