@@ -1,10 +1,5 @@
 using System;
 
-namespace DcmAnonymize
-{
-    public class AnonymizedSeries
-    {
-        public string SeriesInstanceUID { get; set; }
-        public DateTime SeriesDateTime { get; set; }
-    }
-}
+namespace DcmAnonymize.Series;
+
+public record AnonymizedSeries(string SeriesInstanceUID, DateTime SeriesDateTime);
