@@ -3,39 +3,38 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Threading.Tasks;
-using Dicom;
+using FellowOakDicom;
 using KeyedSemaphores;
 
-namespace DcmAnonymize.Series
+namespace DcmAnonymize.Series;
+
+public class SeriesAnonymizer
 {
-    public class SeriesAnonymizer
+    private readonly ConcurrentDictionary<string, AnonymizedSeries> _anonymizedStudies = new ConcurrentDictionary<string, AnonymizedSeries>();
+
+    public async Task AnonymizeAsync(DicomFileMetaInformation metaInfo, DicomDataset dicomDataSet)
     {
-        private readonly ConcurrentDictionary<string, AnonymizedSeries> _anonymizedStudies = new ConcurrentDictionary<string, AnonymizedSeries>();
-
-        public async Task AnonymizeAsync(DicomFileMetaInformation metaInfo, DicomDataset dicomDataSet)
-        {
-            var originalSeriesInstanceUID = dicomDataSet.GetSingleValue<string>(DicomTag.SeriesInstanceUID);
+        var originalSeriesInstanceUID = dicomDataSet.GetSingleValue<string>(DicomTag.SeriesInstanceUID);
     
-            if (!_anonymizedStudies.TryGetValue(originalSeriesInstanceUID, out var anonymizedSeries))
+        if (!_anonymizedStudies.TryGetValue(originalSeriesInstanceUID, out var anonymizedSeries))
+        {
+            using (await KeyedSemaphore.LockAsync($"SERIES_{originalSeriesInstanceUID}"))
             {
-                var key = $"SERIES_{originalSeriesInstanceUID}";
-                using (await KeyedSemaphore.LockAsync(key))
+                if (!_anonymizedStudies.TryGetValue(originalSeriesInstanceUID, out anonymizedSeries))
                 {
-                    if (!_anonymizedStudies.TryGetValue(originalSeriesInstanceUID, out anonymizedSeries))
-                    {
-                        anonymizedSeries = new AnonymizedSeries();
+                    var seriesInstanceUID = DicomUIDGenerator.GenerateDerivedFromUUID().UID;
+                    var seriesDateTime = DateTime.Now;
 
-                        anonymizedSeries.SeriesInstanceUID = DicomUIDGenerator.GenerateDerivedFromUUID().UID;
-                        anonymizedSeries.SeriesDateTime = DateTime.Now;
-
-                        _anonymizedStudies[originalSeriesInstanceUID] = anonymizedSeries;
-                    }
+                    anonymizedSeries = new AnonymizedSeries(seriesInstanceUID, seriesDateTime);
+                        
+                    _anonymizedStudies[originalSeriesInstanceUID] = anonymizedSeries;
                 }
             }
-
-            dicomDataSet.AddOrUpdate(DicomTag.SeriesInstanceUID, anonymizedSeries.SeriesInstanceUID);
-            dicomDataSet.AddOrUpdate(DicomTag.SeriesDate, anonymizedSeries.SeriesDateTime.ToString("yyyyMMdd", CultureInfo.InvariantCulture));
-            dicomDataSet.AddOrUpdate(DicomTag.SeriesTime, anonymizedSeries.SeriesDateTime.ToString("HHmmss", CultureInfo.InvariantCulture));
         }
+
+        dicomDataSet.AddOrUpdate(DicomTag.SeriesInstanceUID, anonymizedSeries.SeriesInstanceUID);
+        dicomDataSet.AddOrUpdate(DicomTag.SeriesDate, anonymizedSeries.SeriesDateTime.ToString("yyyyMMdd", CultureInfo.InvariantCulture));
+        dicomDataSet.AddOrUpdate(DicomTag.SeriesTime, anonymizedSeries.SeriesDateTime.ToString("HHmmss", CultureInfo.InvariantCulture));
+        dicomDataSet.Remove(DicomTag.RelatedSeriesSequence);
     }
 }
