@@ -1,21 +1,14 @@
 using System;
-using System.Diagnostics;
 using DcmAnonymize.Names;
-using Dicom;
 
-namespace DcmAnonymize
-{
-    public class AnonymizedStudy
-    {
-        public string StudyInstanceUID { get; set; }
-        public string Description { get; set; }
-        public string AccessionNumber { get; set; }
-        public DateTime StudyDateTime { get; set; }
-        public RandomName StudyRequestingPhysician { get; set; }
+namespace DcmAnonymize;
 
-        public RandomName StudyPerformingPhysician { get; set; }
-        public string StudyID { get; set; }
-
-        public string InstitutionName { get; set; }
-    }
-}
+public record AnonymizedStudy(
+    string StudyInstanceUID,
+    string AccessionNumber,
+    DateTime StudyDateTime,
+    RandomName RequestingPhysician,
+    RandomName PerformingPhysician,
+    string StudyId,
+    string InstitutionName
+);
